Some review improvements
