+++ src/scrapers/platformen/__init__.py
@@ -1,4 +1,5 @@
 from .scraper import Scraper
+
 from .coronahelpers import CoronaHelpersScraper
 # from .heldnodig import HeldNodig
 # from .mensendiewillenhelpen import MensenDieWillenHelpen
@@ -7,21 +8,14 @@ from .nijmegenoost import NijmegenOost
 from .nlvoorelkaar import NLvoorElkaar
 # from .zorgheldenauto import Zorgheldenauto
 from .puurpapendrecht import PuurPapendrecht
-
 from .wijamsterdam import WijAmsterdam
 
 __all__ = [
     'CoronaHelpersScraper',
-    'NLvoorElkaar',
-    'WijAmsterdam',
-    'Scraper',
     'Maasburen',
     'NijmegenOost',
-    'PuurPapendrecht'
+    'NLvoorElkaar',
+    'PuurPapendrecht',
+    'Scraper',
+    'WijAmsterdam',
 ]
-
-# 'HeldNodig',
-# 'MensenDieWillenHelpen',
-# 'NijmegenOost',
-# 'PuurPapendrecht',
-# 'Zorgheldenauto',
