+++ src/tests/test_helpers.py
@@ -1,4 +1,5 @@
 import random
+from collections import namedtuple
 from datetime import date
 from unittest import TestCase
 
@@ -28,10 +29,10 @@ class TestHelpers(TestCase):
         assert MijnBuurtjeSource.format_organizer('eerste deel') == 'eerste'
 
     def test_format_date_should_raise_on_segment_length(self):
-        for date in ["18", "18 mei", "18 mei 2020 16:00"]:
-            with self.subTest("Error date", date=date):
-                with self.assertRaises(ValueError):
-                    MijnBuurtjeSource.format_date(date)
+        for bad_date in ["18", "18 mei", "18 mei 2020 16:00"]:
+            with self.subTest("Error date", date=bad_date):
+                with self.assertRaisesRegex(ValueError, f"Expecting a three segment date, got {bad_date}."):
+                    MijnBuurtjeSource.format_date(bad_date)
 
     def test_format_months(self):
         for i, month in enumerate(["JANUARI",
@@ -51,3 +52,15 @@ class TestHelpers(TestCase):
                 year = 2000 + random.randint(20, 30)
                 actual = MijnBuurtjeSource.format_date(f"{day} {month} {year}")
                 assert date(year, i + 1, day) == actual
+
+    def test_format_date_should_raise_value_error(self):
+        with self.assertRaises(ValueError):
+            assert None is MijnBuurtjeSource.format_date("")
+
+    def test_should_strip_none_if_none(self):
+        assert None is MijnBuurtjeSource.strip_text(None, "test")
+
+    def test_should_strip_none_if_no_text(self):
+        Element = namedtuple('Element', 'text')
+        assert None is MijnBuurtjeSource.strip_text(Element(text=None), "test")
+
